Adding validation service and finance calculation services, added more maps.

+++ Podium.Data.Primitives/Entities/Product.cs
@@ -9,8 +9,8 @@ namespace Podium.Data.Primitives.Entities
         public string Id { get; set; }
 
         public string Lender { get; set; }
-        public double InterestRate { get; set; }
+        public decimal InterestRate { get; set; }
         public Enums.LoanType LoanType { get; set; }
-        public double LoanToValue { get; set; }
+        public decimal LoanToValue { get; set; }
     }
 }
